refactor: Move Function
This commit removes the permissions for the user_account into the actual user account folder rather then having them separated from the rest of the code base
